+++ src/index.ts
@@ -1,5 +1,5 @@
 import { ExtensionContext } from "@foxglove/studio";
-import { initSystemViewerPanel } from "./SystemViewerPanel2";
+import { initSystemViewerPanel } from "./SystemViewerPanel";
 
 export function activate(extensionContext: ExtensionContext) {
   extensionContext.registerPanel({
