+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
   "extends": "@foxglove/fox/tsconfig/tsconfig.json",
 
-  "include": ["./src/**/*", "src/SystemViewerPanel2.tsx"],
+  "include": ["./src/**/*", "src/SystemViewerPanel.tsx"],
   "compilerOptions": {
     "rootDir": "./src",
     "outDir": "./dist",
